fix  dto && exception && mapper

+++ src/main/java/com/example/service/UserService.java
@@ -1,37 +1,63 @@
 package com.example.service;
 
+import com.example.dto.*;
 import com.example.entity.User;
+import com.example.exception.EmailAlreadyExistsException;
+import com.example.exception.NotFoundException;
+import com.example.mapper.UserMapper;
 import com.example.repository.UserRepository;
-import javax.transaction.Transactional;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
 @Service
 @Transactional
 public class UserService {
-    @Autowired
+
     private final UserRepository repository;
 
-    public UserService(UserRepository repository){
+    @Autowired
+    public UserService(UserRepository repository) {
         this.repository = repository;
     }
 
+    public UserResponseDTO create(UserCreateDTO dto) {
+        if (repository.existsByEmail(dto.getEmail()))
+            throw new EmailAlreadyExistsException("Email already exists");
 
-    public User save(User user){
-        return repository.save(user);
+        User user = UserMapper.toEntity(dto);
+        repository.save(user);
+        return UserMapper.toDto(user);
     }
 
-    public User findById(Long id) {
-        return repository.findById(id).orElse(null);
-    }
-
-    public List<User> findAll(){
+    public List<User> findAll() {
         return repository.findAll();
     }
 
-    public void deleteById(Long id){
+    public UserResponseDTO findById(Long id) {
+        User user = repository.findById(id)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+        return UserMapper.toDto(user);
+    }
+
+    public UserResponseDTO update(Long id, UserUpdateDTO dto) {
+        User user = repository.findById(id)
+                .orElseThrow(() -> new NotFoundException("User not found"));
+        user.setName(dto.getName());
+        user.setEmail(dto.getEmail());
+        user.setPassword(dto.getPassword());
+        user.setRole(dto.getRole());
+        user.setActive(dto.getActive());
+
+        repository.save(user);
+        return UserMapper.toDto(user);
+    }
+
+    public void delete(Long id) {
+        if (!repository.existsById(id))
+            throw new NotFoundException("User not found");
         repository.deleteById(id);
     }
 }
